requirements.txt
replicate
pillow
numpy
torch
google-genai
opencv-python
torchaudio
av